refactor: decompose CppAnalyzer monolith and resolve circular dependencies
This change implements a significant architectural refactoring to improve modularity and maintainability:
- Decomposed the ~6,000 line CppAnalyzer into cohesive components:
  - ClangParser: Handles libclang Index and TranslationUnit lifecycle.
  - SymbolExtractor: Manages AST traversal and metadata extraction.
  - WorkerPoolManager: Encapsulates concurrency and process/thread pool logic.
- Resolved the circular dependency between cpp_mcp_server and consolidated_tools using a new ToolRegistry.
- Added facade methods to CppAnalyzer to maintain backward compatibility with existing tests and internal callers.
- Updated regression tests to work with the new decoupled architecture.
