+++ tests/test_usr_to_display_name.py
@@ -6,7 +6,7 @@ human-readable qualified names, preserving template arguments.
 
 import pytest
 
-from mcp_server.cpp_analyzer import (
+from mcp_server.symbol_extractor import (
     _decode_class_ref,
     _decode_template_args,
     _decode_usr_type,
